добавляет в localstorage watched and queue
